Refactored Serializers
