+++ src/apps/api/serializers/tasks.py
@@ -4,6 +4,7 @@ from rest_framework.exceptions import ValidationError
 
 from api.mixins import DefaultUserCreateMixin
 from api.serializers.datasets import DataDetailSerializer, DataSimpleSerializer
+from competitions.models import PhaseTaskInstance, Phase
 from datasets.models import Data
 from tasks.models import Task, Solution
 
@@ -125,3 +126,32 @@ class TaskListSerializer(serializers.ModelSerializer):
             # Value is used for Semantic Multiselect dropdown api calls
             'value',
         )
+
+class PhaseTaskInstanceSerializer(serializers.HyperlinkedModelSerializer):
+    task = serializers.SlugRelatedField(queryset=Task.objects.all(), required=True, allow_null=False, slug_field='key',
+                                        many=False)
+    phase = serializers.PrimaryKeyRelatedField(many=False, queryset=Phase.objects.all())
+    id = serializers.IntegerField(source='task.id')
+    value = serializers.CharField(source='task.key', required=False)
+    key = serializers.CharField(source='task.key', required=False)
+    created_when = serializers.DateTimeField(source='task.created_when')
+    name = serializers.CharField(source='task.name')
+
+
+    class Meta:
+        model = PhaseTaskInstance
+        fields = (
+            'task',
+            'order_index',
+            'phase',
+            'id',
+            # Value is used for Semantic Multiselect dropdown api calls
+            'value',
+            'key',
+            'created_when',
+            'name',
+            # 'solutions',
+            # 'ingestion_only_during_scoring',
+        )
+
+
