+++ src/apps/api/views/competitions.py
@@ -110,8 +110,6 @@ class CompetitionViewSet(ModelViewSet):
             return LeaderboardPhaseSerializer
         elif self.request.method == 'GET':
             return CompetitionDetailSerializer
-        # elif self.request.method == 'PATCH':
-        #     return CompetitionUpdateSerializer
         else:
             return CompetitionSerializer
 
@@ -133,8 +131,17 @@ class CompetitionViewSet(ModelViewSet):
         in the response to remove a GET from the frontend"""
         partial = kwargs.pop('partial', False)
         instance = self.get_object()
+
+        from pprint import pprint
+        print(f'\n\n\nREQUEST DATA\n@@@@@@@@@@@@@@@@@@@@@@@@@@@@@`')
+        pprint(request.data)
+
         serializer = self.get_serializer(instance, data=request.data, partial=partial)
         serializer.is_valid(raise_exception=True)
+
+        print(f'\n\n\nSERIALIZER DATA\n@@@@@@@@@@@@@@@@@@@@@@@@@@@@@`')
+        pprint(serializer.data)
+        print('\nEND\n\n\n')
         self.perform_update(serializer)
 
         if getattr(instance, '_prefetched_objects_cache', None):
